# generate zipped goto and action tables