As discussed in issue https://github.com/goccmack/gocc/issues/28

I used the katydid repo to benchmark this change, since it has very big tables
https://github.com/katydid/katydid

I only tested using go 1.7 I didn't try the new go 1.8 changes that are mentioned in the issue.

```
$ go version
go version go1.7 darwin/amd64
```

To reproduce clone katydid and go into the relapse/bnf folder

```
$ cd relapse/bnf/
```

Current generated code

```
$ time gocc -o .. all.bnf

real    0m24.734s
user    0m40.693s
sys 0m0.211s

$ wc -l ../parser/gototable.go
  119038 ../parser/gototable.go

$ wc -l ../parser/actiontable.go
  105239 ../parser/actiontable.go

$ time gofmt -l -s -w ../parser/
../parser/action.go
../parser/actiontable.go
../parser/gototable.go
../parser/parser.go
../parser/productionstable.go

real    0m1.539s
user    0m1.812s
sys 0m0.043s

$ time go install ../parser/

real    0m5.805s
user    0m7.608s
sys 0m0.359s
```

New generated code

```
$ time gocc -zip -o .. all.bnf

real    0m23.699s
user    0m39.755s
sys 0m0.181s

$ wc -l ../parser/gototable.go
  433 ../parser/gototable.go

$ wc -l ../parser/actiontable.go
  1092 ../parser/actiontable.go

$ time gofmt -l -s -w ../parser/
../parser/action.go
../parser/actiontable.go
../parser/gototable.go
../parser/parser.go
../parser/productionstable.go

real    0m0.094s
user    0m0.092s
sys 0m0.009s

$ time go install ../parser/

real    0m0.969s
user    0m1.335s
sys 0m0.124s
```

Code generation time is about the same.

The amount of generated code is about 100 - 200 times less.

gofmt is about a 100 times faster, but this is still small compared to the generation time.

The go install time is about 5 times faster.