@awalterschulze 2016-10-30
There is still some cleaning up to do, but I think we should first decide if this is something we want.
I think just having code base so much less influenced by having a 100 times less generated code might be worth it.
What do you @mewmew @goccmack think?