@awalterschulze 2016-12-01
Ok I decided I really want this.
Do you have any objections if I merge this?
@goccmack @mewmew