@mewmew 2016-12-01
@awalterschulze, Hej! Sorry for the delay. I finally had time to take a deeper look. Really surprised you managed to achieve this in only ~150 lines of code. Looks clean to use gzip and gob, with hex output. This is a great time saver for development, and when something looks strange, I shift back to regular output with debug enabled. I am glad this is not the default, but guarded by the `-zip` flag. Still think we may mark the `-zip` flag as experimental, but definitely add it to gocc so we can use it and further stress test it. If we decide to revisit this in the future, and evaluate against a proper implementation of Pager's "practical general method", we can evaluate both side by side.

For now, I must say, I would also very much like to start using this, as it shaves of roughly 30% of the time from tasks I run 50x times a day :) Thank you very much for pushing this forward.

Here are some results from using your zip implementation to generate the lexer and parser of the [LLVM IR library](https://github.com/llir/llvm/blob/parser/asm/internal/ll.bnf) which is currently under heavy development.

# Lines of code

## Before

```
u@x1 ~/D/g/s/g/l/l/a/internal> cloc .
      47 text files.
      47 unique files.                              
      17 files ignored.

github.com/AlDanial/cloc v 1.70  T=0.97 s (31.1 files/s, 204530.4 lines/s)
-------------------------------------------------------------------------------
Language                     files          blank        comment           code
-------------------------------------------------------------------------------
Go                              17           4796           1581         190790
C                               10             31              0            142
make                             2             11              1             53
Bourne Shell                     1              0              0             26
-------------------------------------------------------------------------------
SUM:                            30           4838           1582         191011
-------------------------------------------------------------------------------
```

## After

```
u@x1 ~/D/g/s/g/l/l/a/internal> cloc .
      47 text files.
      47 unique files.                              
      17 files ignored.

github.com/AlDanial/cloc v 1.70  T=0.16 s (183.9 files/s, 165304.5 lines/s)
-------------------------------------------------------------------------------
Language                     files          blank        comment           code
-------------------------------------------------------------------------------
Go                              17           3175           1581          21941
C                               10             31              0            142
make                             2             11              1             53
Bourne Shell                     1              0              0             26
-------------------------------------------------------------------------------
SUM:                            30           3217           1582          22162
-------------------------------------------------------------------------------
```

# Lexer and parser generation time

## Before

```
u@x1 ~/D/g/s/g/l/l/a/internal> time gocc ll.bnf
real 2.16
user 2.89
sys 0.02
```

## After

```
u@x1 ~/D/g/s/g/l/l/a/internal> time gocc -zip ll.bnf
real 2.02
user 2.66
sys 0.02
```

# Compilation time

## Before

```
u@x1 ~/D/g/s/g/l/l/a/internal> time go install -a ./...
real 9.62
user 15.41
sys 0.66
```

## After

```
u@x1 ~/D/g/s/g/l/l/a/internal> time go install -a ./...
real 6.00
user 12.08
sys 0.53
```

# Testing time

## Before

```
u@x1 ~/D/g/s/g/l/l/a/internal> time go test ./...
?       github.com/llir/llvm/asm/internal/errors    [no test files]
ok      github.com/llir/llvm/asm/internal/irx   0.705s
?       github.com/llir/llvm/asm/internal/lexer [no test files]
ok      github.com/llir/llvm/asm/internal/parser    0.006s
?       github.com/llir/llvm/asm/internal/token [no test files]
?       github.com/llir/llvm/asm/internal/util  [no test files]
real 5.67
user 7.94
sys 0.37
```

## After

```
u@x1 ~/D/g/s/g/l/l/a/internal> time go test ./...
?       github.com/llir/llvm/asm/internal/errors    [no test files]
ok      github.com/llir/llvm/asm/internal/irx   0.703s
?       github.com/llir/llvm/asm/internal/lexer [no test files]
ok      github.com/llir/llvm/asm/internal/parser    0.011s
?       github.com/llir/llvm/asm/internal/token [no test files]
?       github.com/llir/llvm/asm/internal/util  [no test files]
real 1.87
user 3.30
sys 0.14
```

# Summary

## Lines of code

88% less lines of code.

## Lexer and parser generation time

6% faster generation.

## Compilation time

37% faster compilation.

## Testing time

32% faster testing.

So from me, you have a +1 for sure to merge this! Great work.