@awalterschulze 2016-12-02
Ok awesome.

I'll see if I can clean it up and merge it asap.