@awalterschulze 2016-12-02
Ok I guess its clean enough :)