@awalterschulze 2016-12-02
Yeah just deleted 2/3s of my project or over 220 000 lines of code
https://github.com/katydid/katydid/commit/2c0e1b81d533a9cc7e350e1addaf6e1e401f4849